DOPE-237: changed set operators to enums for slimmer code
