+++ core/src/test/kotlin/ch/ergon/dope/buildTest/QueryBuilderTest.kt
@@ -9,12 +9,12 @@ import ch.ergon.dope.helper.someFromClause
 import ch.ergon.dope.helper.someNumberField
 import ch.ergon.dope.helper.someStringField
 import ch.ergon.dope.helper.unifyString
-import ch.ergon.dope.resolvable.clause.model.setoperators.except
-import ch.ergon.dope.resolvable.clause.model.setoperators.exceptAll
-import ch.ergon.dope.resolvable.clause.model.setoperators.intersect
-import ch.ergon.dope.resolvable.clause.model.setoperators.intersectAll
-import ch.ergon.dope.resolvable.clause.model.setoperators.union
-import ch.ergon.dope.resolvable.clause.model.setoperators.unionAll
+import ch.ergon.dope.resolvable.clause.model.setoperator.except
+import ch.ergon.dope.resolvable.clause.model.setoperator.exceptAll
+import ch.ergon.dope.resolvable.clause.model.setoperator.intersect
+import ch.ergon.dope.resolvable.clause.model.setoperator.intersectAll
+import ch.ergon.dope.resolvable.clause.model.setoperator.union
+import ch.ergon.dope.resolvable.clause.model.setoperator.unionAll
 import ch.ergon.dope.resolvable.expression.alias
 import ch.ergon.dope.resolvable.expression.unaliased.type.FALSE
 import ch.ergon.dope.resolvable.expression.unaliased.type.MISSING
